% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.mfilter.R
\name{autoplot.mFilter}
\alias{autoplot.mFilter}
\title{autoplot for mFilter}
\usage{
\method{autoplot}{mFilter}(x)
}
\arguments{
\item{x}{mFilter class object}
}
\description{
Add mFilter to autplot
}
\seealso{
\code{\link[mFilter]{mFilter}}
}
