% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tseries.plot.R
\docType{package}
\name{tseries.plot-package}
\alias{tseries.plot}
\alias{tseries.plot-package}
\title{Time Series Plot}
\description{
Plots for time series object based in ggplot2 package
and print tables with basic statistical values.
}
\author{
Victor Espinoza
}
