% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.break.coef.R
\name{tsp.break.coef}
\alias{tsp.break.coef}
\title{Break points coefficients and t values}
\usage{
tsp.break.coef(
  data,
  breaks = NULL,
  model = c("trend", "const"),
  log = FALSE,
  plot = TRUE,
  mfcol = NULL,
  ...
)
}
\arguments{
\item{data}{an object of class ts or as.ts() supported}

\item{breaks}{numbers of maximum breaks to found}

\item{model}{type of model to be used in breakpoints, trend (default) or const (for constant only)}

\item{log}{log transformation used in data for log percent change}

\item{plot}{logical, plot breaks using graphics}

\item{mfcol}{vector of length  2 to define plot matrix}

\item{...}{arguments passed to breakpoints}
}
\value{
tibble object
}
\description{
Break points coefficients and t values
}
\details{
Create a linear model to determinate structural breaks
}
\examples{
library(vars)
data("Canada")
tsp.break.coef(Canada)

}
