% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.break.R
\name{tsp.break}
\alias{tsp.break}
\title{Breakpoints plot for time series}
\usage{
tsp.break(model, level = NULL, ...)
}
\arguments{
\item{model}{breakpointsfull or breakpoints class object created with strucchange::breakpoints()}

\item{level}{confidence level used in stats::confint()}

\item{...}{additional argument(s) for stats::confint()}
}
\value{

}
\description{
Breakpoints plot for time series
}
