% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.correlogram.R
\name{tsp.correlogram}
\alias{tsp.correlogram}
\title{correlogram plot from a time series}
\usage{
tsp.correlogram(
  x,
  type = c("Ljung-Box", "Box-Pierce"),
  lag.max = NA,
  table = TRUE,
  seasonal = FALSE
)
}
\arguments{
\item{x}{ts or Arima object}

\item{type}{type of portmanteau test to use.
"Ljung-Box" (default) and "Box-Pierce" are allowed.}

\item{lag.max}{lag max used in acf, pacf and portmanteau test.
If lag.max is NA (default), lag.max = 10 for non seasonal ts
or Arima residual and 2*frequency(m) for seasonal ts or Arima residual.}

\item{table}{logical. if TRUE (default) a table is displayed}

\item{seasonal}{if TRUE (default) acf and pacf plot
show seasonal lag color different from regular lag}
}
\description{
Plot Acf, Pacf and Portmanteau test for ts or arima class object
}
\details{
plot Acf, Pacf and portmanteau test for
a time series object (ts class) or Arima residual
(Arima class) created with Arima() or arima(). If x
is Arima class object, residual are used to plot, and
fitdf inside portmanteau test is fitted to: lag-length(x$coef).
If table = TRUE, a table with number of lag, acf and pacf values
is displayed, with Q stat, degree of freedom and p-value for
portmanteau test.
}
\seealso{
\code{\link[stats]{ts}}
\code{\link[stats]{arima}}
\code{\link[stats]{Box.test}}
\code{\link[forecast]{Arima}}
\code{\link[forecast]{Acf}}
\code{\link[forecast]{Pacf}}
\code{\link[forecast]{checkresiduals}}
}
\author{
Víctor Espinoza
}
