% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.granger.R
\name{tsp.granger}
\alias{tsp.granger}
\title{Granger Causality test for multivariate VAR models}
\usage{
tsp.granger(model)
}
\arguments{
\item{model}{VAR object created with vars::VAR()}
}
\value{
tibble class data with causality test for each pair
}
\description{
Perform Granger causality block test based on Wald Chi-Squared
}
\examples{
library(tseries.plot)
library(vars)
data("Canada")
tsp.granger(VAR(Canada,2))
}
