% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.mfilter.R
\name{tsp.mfilter}
\alias{tsp.mfilter}
\title{autoplot for mFilter}
\usage{
tsp.mfilter(x)
}
\arguments{
\item{x}{mFilter class object}
}
\description{
Add mFilter to autplot
}
\seealso{
\code{\link[mFilter]{mFilter}}
}
