% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.nominal.R
\name{tsp.nominal}
\alias{tsp.nominal}
\title{Time series with nominal variable}
\usage{
tsp.nominal(series, nominal, ...)
}
\arguments{
\item{series}{univariable timeseries object of class ts}

\item{nominal}{dichotomous(binary) vector to use as categorical variable}

\item{...}{arguments to be passed to plot()}
}
\value{
plot object
}
\description{
Time series with nominal variable
}
\examples{
series <- rnorm(120,120,120)
nominal <- rbinom(120,1,0.5)
tsp.nominal(series,nominal)

}
