% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.range.mean.R
\name{tsp.range.mean}
\alias{tsp.range.mean}
\title{Range Mean Plot}
\usage{
tsp.range.mean(x, table = T)
}
\arguments{
\item{x}{ts class object}

\item{table}{logical, if true (default) print mean and range values}
}
\description{
Plot a Range Mean Plot using ggplot2 package
}
\seealso{
\code{\link[base]{mean}}
\code{\link[base]{max}}
\code{\link[base]{min}}
}
