% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.range_mean.R
\name{tsp.range_mean}
\alias{tsp.range_mean}
\title{Range mean plot for time series}
\usage{
tsp.range_mean(x)
}
\arguments{
\item{x}{univariate time series object, created with ts()}
}
\value{
table and plot with range mean information
}
\description{
Range mean plot for time series
}
\examples{
library(tseries.plot)
tsp.range_mean(AirPassengers)

}
