% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.season.box.R
\name{tsp.season.box}
\alias{tsp.season.box}
\title{Season plot using boxplot}
\usage{
tsp.season.box(x = x, type = c("none", "multiplicative", "additive"),
  table = TRUE)
}
\arguments{
\item{x}{time series object of class "ts"}

\item{type}{decomposition type multiplicative (default)
or additive.}

\item{table}{logical, if true(default) median, mean, sd and var
values are printed, using season as group}
}
\description{
Plot using ggplot2 a
time series object, ts class. Season is
used as factor to group also use
decompose method to remove trend
}
