% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.season.line.R
\name{tsp.season.line}
\alias{tsp.season.line}
\title{Season plot by line}
\usage{
tsp.season.line(x)
}
\arguments{
\item{x}{time series to plot,
only ts class}
}
\description{
Plot using ggplot2 a
time series using year as group
}
