% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.series_factor.R
\name{tsp.series_factor}
\alias{tsp.series_factor}
\title{Time series with factor variable plot}
\usage{
tsp.series_factor(x, factor, color = TRUE, gg = TRUE)
}
\arguments{
\item{x}{Univaritate Time series, "ts" class object}

\item{factor}{"factor" class object}

\item{color}{logical, True (default) print plot with colors}

\item{ggplot}{logial, True (default) use ggplot2 device}
}
\value{

}
\description{
Time series with factor variable plot
}
\examples{
series <- ts(rnorm(120,120,120),start = c(2012,1),frequency = 12)
factor <- factor(x = rbinom(120,1,0.9),levels = 0:1,labels = c("No crisis","Crisis"))
tsp.series_factor(series,factor)
}
