% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.ty_causality.R
\name{tsp.toda_yamamoto}
\alias{tsp.toda_yamamoto}
\title{Toda and Yamamoto Causality Test}
\usage{
tsp.toda_yamamoto(data, p = 1, d = 1, print_model = FALSE, ...)
}
\arguments{
\item{data}{object of class mts created with ts()}

\item{p}{lag order for VAR, use vars::VARselect()}

\item{d}{max integration order I(d), use Unit Root test to determinate}

\item{print_model}{Logical, print VAR model used for test}

\item{...}{extra parameters used in vars::VAR()}
}
\value{
list object with VAR and Toda-Yamamoto causality test
}
\description{
Toda and Yamamoto Causality Test
}
\examples{
library(tseries.plot)
library(vars)
data("Canada")
tsp.toda_yamamoto(Canada,1,1)
}
