% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.trend.decompose.R
\name{tsp.trend.decompose}
\alias{tsp.trend.decompose}
\title{Trend Decomposed Plot}
\usage{
tsp.trend.decompose(x, type = c("multiplicative", "additive"))
}
\arguments{
\item{x}{time series object of class "ts"}

\item{type}{decomposition type multiplicative (default)
or additive.}
}
\description{
Use the base decompose function to
plot a trend from a time series object using ggplot2
}
\seealso{
\code{\link[stats]{decompose}}
}
\author{
Victor Espinoza
}
