% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.var.fit.R
\name{tsp.var.fit}
\alias{tsp.var.fit}
\title{Fitted and observed values from VAR model}
\usage{
tsp.var.fit(model)
}
\arguments{
\item{model}{'varest' class object from vars::VAR()}
}
\value{
grobs object from gridExtra::grid.arrange()
}
\description{
Fitted and observed values from VAR model
}
