% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.var.forecast.R
\name{tsp.var.forecast}
\alias{tsp.var.forecast}
\title{Predicts and plot VAR model}
\usage{
tsp.var.forecast(model, n_ahead = 12, color = NULL, n_col = NULL, ...)
}
\arguments{
\item{model}{VAR class objects from vars::VAR()}

\item{n_ahead}{integer, number of periods to forecast}

\item{color}{color used for min and max values in forecast}

\item{n_col}{number of columns used in plot}

\item{...}{arguments passed to vars::predict()}
}
\value{
plot from gridExtra::grid.arrange()
}
\description{
Predicts and plot VAR model
}
\examples{
library(vars)
library(tseries.plot)
data("Canada")
model <- VAR(Canada,3)
tsp.var.forecast(model,n_ahead = 4,color = "blue",n_col = 2)

}
