% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.var.irf.R
\name{tsp.var.irf}
\alias{tsp.var.irf}
\title{Irf VAR plot}
\usage{
tsp.var.irf(x)
}
\arguments{
\item{x}{irf class object created by vars::irf() function}
}
\value{
A ggplot2 object class
}
\description{
Irf VAR plot
}
