% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.var.resid.R
\name{tsp.var.resid}
\alias{tsp.var.resid}
\title{Residuals plot from VAR model}
\usage{
tsp.var.resid(model)
}
\arguments{
\item{model}{'varest' class object from vars::VAR()}
}
\value{
grobs object from gridExtra::grid.arrange()
}
\description{
Residuals plot from VAR model
}
