% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.year.R
\name{tsp.year}
\alias{tsp.year}
\title{time series plot by year}
\usage{
tsp.year(x = x, table = TRUE)
}
\arguments{
\item{x}{a time series object (ts class)}

\item{table}{logical, if TRUE (default) print a table with year,
median, mean, sd and variance values}
}
\description{
Draw a boxplot of a time series, using year as grouping factor
}
\details{
Draw a boxplot for a time series object created with ts()
using year as grouping factor.
}
\seealso{
\code{\link[stats]{ts}}
\code{\link[base]{trunc.Date}}
\code{\link[stats]{time}}
}
\author{
Víctor Espinoza
}
