% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.season.line.R
\name{tsp.season.line}
\alias{tsp.season.line}
\title{Season plot by line}
\usage{
tsp.season.line(x)
}
\arguments{
\item{x}{time series, ts class object
only ts class}
}
\description{
Line plot of time series using
year as group
}
