% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.ty_causality.R
\name{tsp.toda_yamamoto}
\alias{tsp.toda_yamamoto}
\title{Toda and Yamamoto Causality Test}
\usage{
tsp.toda_yamamoto(
  data,
  p = NULL,
  d = NULL,
  criteria = c("aic", "hq", "sc", "fpe"),
  test = c("adf", "kpss", "pp"),
  ...
)
}
\arguments{
\item{data}{object of class mts created with ts()}

\item{p}{lag order for VAR, use vars::VARselect(), if NULL (default) automatic value is calculated}

\item{d}{max integration order I(d), if NULL (default) automatic value is calculated}

\item{criteria}{information criteria for automatic p order, used in vars::VARselect()}

\item{test}{Type of unit root test to use for automatic d value, used in forecast::ndiffs()}

\item{...}{extra parameters used in vars::VAR()}
}
\value{
list object with VAR and Toda-Yamamoto causality test
}
\description{
Toda and Yamamoto Causality Test
}
\examples{
library(tseries.plot)
library(vars)
data("Canada")
tsp.toda_yamamoto(Canada,1,1,type="both")
}
