% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.var.fevd.R
\name{tsp.var.fevd}
\alias{tsp.var.fevd}
\title{FEVD plot on ggplot2}
\usage{
tsp.var.fevd(fevd)
}
\arguments{
\item{x}{A fevd class object created by vars::fevd()}
}
\value{
a ggplot2 object class
}
\description{
FEVD plot on ggplot2
}
\examples{

}
