% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.var.irf.R
\name{tsp.var.irf}
\alias{tsp.var.irf}
\title{Irf VAR plot}
\usage{
tsp.var.irf(irf, color = NULL)
}
\arguments{
\item{irf}{irf class object created by vars::irf() function}

\item{color}{color fill for ggplot2::geom_ribbon()}
}
\value{
A ggplot2 object class with irf from vars::irf()
}
\description{
Irf VAR plot
}
